refactor: printing error message if the ID is not found

+++ src/main/java/com/pointdafamilia/pointdafamilia/Services/BebidasService.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 import com.pointdafamilia.pointdafamilia.dtos.BebidasDTO;
 import com.pointdafamilia.pointdafamilia.dtos.BebidasDTOAdmin;
 import com.pointdafamilia.pointdafamilia.entities.Bebidas;
+import com.pointdafamilia.pointdafamilia.exceptions.BebidaNotFound;
 import com.pointdafamilia.pointdafamilia.repositories.BebidasRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -31,18 +32,26 @@ public class BebidasService {
     }
 
     public BebidasDTO buscarPorId(Long id) {
-        Bebidas bebida = bebidasRepository.findById(id).orElseThrow();
+        Bebidas bebida = bebidasRepository.findById(id).orElseThrow(()-> new BebidaNotFound(id));
         return modelMapper.map(bebida, BebidasDTO.class);
     }
 
     public BebidasDTO atualizarBebidas(Long id, BebidasDTO bebidasDTO) {
-        Bebidas bebida = modelMapper.map(bebidasDTO, Bebidas.class);
-        bebida.setId(id);
-        bebida = bebidasRepository.save(bebida);
+        Bebidas bebida = bebidasRepository.findById(id).orElseThrow(()-> new BebidaNotFound(id));
+        bebida.setNome(bebidasDTO.getNome());
+        bebida.setPreco(bebidasDTO.getPreco());
+        bebida.setTipo(bebidasDTO.getTipo());
+        bebida.setImagem(bebidasDTO.getImagem());
+        bebida.setCategoria(bebidasDTO.getCategoria());
+        bebidasRepository.save(bebida);
+
         return modelMapper.map(bebida, BebidasDTO.class);
     }
 
     public void deletarBebida(Long id) {
+        if (!bebidasRepository.existsById(id)) {
+            throw new BebidaNotFound(id);
+        }
         bebidasRepository.deleteById(id);
     }
